+++ R/preprocess.R
@@ -1,4 +1,4 @@
-preprocess <- function(X, center = FALSE, scale = FALSE, impute = FALSE) {
+preprocess <- function(X, center = FALSE, scale = FALSE, impute = FALSE, nCores = getOption("mc.cores", 2L)) {
     if (!(is.numeric(X) && length(dim(X)) == 2)) {
         stop("'X' needs to be a numeric matrix")
     }
@@ -11,5 +11,8 @@ preprocess <- function(X, center = FALSE, scale = FALSE, impute = FALSE) {
     if (!(is.logical(impute) && length(impute) == 1L)) {
         stop("'impute' needs to be a logical vector of size 1")
     }
-    .Call(C_preprocess, X, center, scale, impute)
+    if (!(is.numeric(nCores) && nCores > 0L)) {
+        stop("'nCores' needs to be a positive number")
+    }
+    .Call(C_preprocess, X, center, scale, impute, as.integer(nCores))
 }