+++ src/init.c
@@ -7,7 +7,7 @@
 static const R_CallMethodDef callMethods[] = {
     {"summarize", (DL_FUNC) &summarize, 1},
     {"rayOLS", (DL_FUNC) &rayOLS, 2},
-    {"preprocess", (DL_FUNC) &preprocess, 4},
+    {"preprocess", (DL_FUNC) &preprocess, 5},
     {NULL, NULL, 0}
 };