+++ src/preprocess.h
@@ -2,4 +2,4 @@
 
 #include <Rinternals.h>
 
-SEXP preprocess(SEXP sIn, SEXP sCenter, SEXP sScale, SEXP sImpute);
+SEXP preprocess(SEXP sIn, SEXP sCenter, SEXP sScale, SEXP sImpute, SEXP sNumCores);